Add Astlib to replace the use of omp
This commit adds the new module Astlib to ppxlib. Astlib on one hand
includes the various AST versions and conversions between them; they used
to be in the external library OMP. On the other hand, Astlib provides a
minimal API of everything that's used inside ppxlib from the compiler and
ppxlib only uses compiler-libs through Astlib now. To make sure of that,
this commit also disables the use of implicit transitive dependencies
in the dune project.

The final goal is to get a copy of the part of Astlib, that needs to be
maintained, merged into the compiler. Astlib in the compiler will
1. already contain the AST of the new compiler version and the conversions
between it and the older versions as soon as a new compiler gets released
2. maintain the Astlib API stable - and with that, anything used in
ppxlib from the compiler

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>